🚚 rename: 도메인 통합 및 구조 위치 변경
domain/race 내부에 각각이 클래스로 위치했었는데 이를 controller, model, service, util로 폴더를 구분하고 각각에 클래스를 위치시킴
