+++ src/main/java/Application.java
@@ -1,5 +1,9 @@
+import java.util.*;
+
 public class Application {
     public static void main(String[] args) {
+        Scanner sc = new Scanner(System.in);
+
 
     }
 }
