+++ src/main/java/domain/race/util/CarUtils.java
@@ -1,4 +1,4 @@
-package domain.car;
+package domain.race.util;
 
 import java.util.Random;
